iCal feed: Escape text, add name, fix multiday events.
Newlines are replaced by \n, and backslashes, commas and semicolons are
escaped with a backslash.

Add a `add_day` template filter to add one day to the end date of an
event to fix display of multiday events.

Add `CALENDAR_NAME` and `CALENDAR_DESCRIPTION` to settings which are
included as X-WR-CALNAME and X-WR-CALDESC.