Apply cheeky optimization to /tutoradmin/
Executing the QuerySet of ModelChoiceFields' choices and setting itself
to it makes the query only run once and not for every render. This seems
to half the loading time of the page. It's still super slow... This is
due to the amount of HTML that is being generated which takes some time
for the browser.